# Start using Boyer-Moore-Horspool algorithm.