- Changes on `firstIndex` to use Boyer-Moore-Horspool algorithm [swift-algorithm-club](https://github.com/raywenderlich/swift-algorithm-club/tree/master/Boyer-Moore-Horspool) used as a reference.