@benlangmuir 2019-01-11
Thanks for taking a look!  Before accepting this it will need performance tests demonstrating the impact, and the data should be chosen based on where we're using this function.  I already made one attempt to replace this back when I originally wrote this code and my implementation turned out to be slower in practice.  It's been a while, but I think part of the issue is that the callers of this code are searching for short strings.