@benlangmuir 2019-01-11
I checked: the only place this is currently used (aside from test code) is searching for the two byte sequence "\r\n" in message headers.  Before writing your own perf tests you could also sanity check with the existing `ConnectionPerfTests`.  They don't specifically stress the message parsing, but if there's a big change it might show up there.