@abdullahselek 2019-01-21
Below you can find the results for `ConnectionPerfTests` with two different implementations.
**Old**
- `testEcho1` 0,00319s
- `testEcho100Latency` 0,0718s
- `testEcho100Throughput` 0,0253s

**New**
- `testEcho1` 0,00153s
- `testEcho100Latency` 0,066s
- `testEcho100Throughput` 0,0265s