@benlangmuir 2019-01-22
Nice!  I'd still like to see a perf test that's specific to this function before we merge.