@abdullahselek 2019-01-25
I added performance tests for collection, have a look when you have time.